When deleting a student, also delete its profile image
When a client deletes a student, we optimistically delete the profile
image from the './images' folder as well. If we fail to delete the
profile image, we should create a job to keep track of the image so that
we can come back and clean it up later. This section of the code is
currently marked with a TODO comment.